Refactor some test helper functions for converting between floats and the bit representation into a single header file (float_helpers.h) in some of the correctness tests.
