+++ test/correctness/runtime_float16_t_downcast.cpp
@@ -4,6 +4,7 @@
 #include <stdlib.h>
 #include <limits>
 #include <vector>
+#include "float_helpers.h"
 
 // FIXME: We should use a proper framework for this. See issue #898
 void h_assert(bool condition, const char *msg) {
@@ -13,23 +14,6 @@ void h_assert(bool condition, const char *msg) {
     }
 }
 
-float float_from_bits(uint32_t bits) {
-    union {
-        float asFloat;
-        uint32_t asUInt;
-    } out;
-    out.asUInt = bits;
-    return out.asFloat;
-}
-
-double double_from_bits(uint64_t bits) {
-    union {
-        double asDouble;
-        uint64_t asUInt;
-    } out;
-    out.asUInt = bits;
-    return out.asDouble;
-}
 struct Result {
     uint16_t RZ; // Result for round to Zero
     uint16_t RU; // Result for round to +ve infinity
